Fix Get Started Modal, WIP Netlify Request Demo Form
