+++ src/components/Layout/Layout.scss
@@ -1026,7 +1026,6 @@ a.middle {
     position: relative;
     left: 0;
     right: 0;
-    margin-top: 50px;
 }
 
 .docsPageLayout {
