+++ src/pages/index.js
@@ -28,6 +28,44 @@ import { FeaturedSectionTripleImage } from '../components/Sections/FeaturedSecti
 import { Spacer } from '../components/Spacer'
 import { DesignedForYourStackBlock } from '../components/Sections/DesignedForYourStackBlock'
 
+const GetStartedModalContent = () => (
+    <>
+        <h2>Try PostHog for free</h2>
+        <div className="modalCardsWrapper">
+            <a href="https://app.posthog.com/signup">
+                <div className="modalSaasCloud modalCard">
+                    <div className="modalCardHeader">
+                        <img src={modalSaasCloud} alt="modal-saas-cloud" />
+                        <h2>Cloud</h2>
+                    </div>
+                    <h4>Small business or low volumes and don't want hassle?</h4>
+                    <p>This is the simplest way to get started. Create an account.</p>
+                </div>
+            </a>
+            <Link to="/docs/deployment">
+                <div className="modalSelfDeploy modalCard">
+                    <div className="modalCardHeader">
+                        <img src={modalSelfDeploy} alt="modal-self-deploy " />
+                        <h2>Open Source</h2>
+                    </div>
+                    <h4>Want to use our free open source product?</h4>
+                    <p>Deploy PostHog Open Source on your own infrastructure. Free forever.</p>
+                </div>
+            </Link>
+            <a href="mailto:sales@posthog.com">
+                <div className="modalSelfDeploy modalCard">
+                    <div className="modalCardHeader">
+                        <img src={modalSelfDeploy} alt="modal-self-deploy " />
+                        <h2>Enterprise</h2>
+                    </div>
+                    <h4>10K+ users? Need support?</h4>
+                    <p>Managed on your infrastructure with greater scalability and support. Book a pilot.</p>
+                </div>
+            </a>
+        </div>
+    </>
+)
+
 function IndexPage() {
     const [modalIsOpen, setModalIsOpen] = useState(false)
 
@@ -68,47 +106,7 @@ function IndexPage() {
                                             className="modalContent"
                                             overlayClassName="modalOverlay"
                                         >
-                                            <h2>Try PostHog for free</h2>
-                                            <div className="modalCardsWrapper">
-                                                <a href="https://app.posthog.com/signup">
-                                                    <div className="modalSaasCloud modalCard">
-                                                        <div className="modalCardHeader">
-                                                            <img src={modalSaasCloud} alt="modal-saas-cloud" />
-                                                            <h2>Cloud</h2>
-                                                        </div>
-                                                        <h4>Small business or low volumes and don't want hassle?</h4>
-                                                        <p>
-                                                            This is the simplest way to get started. Create an account.
-                                                        </p>
-                                                    </div>
-                                                </a>
-                                                <Link to="/docs/deployment">
-                                                    <div className="modalSelfDeploy modalCard">
-                                                        <div className="modalCardHeader">
-                                                            <img src={modalSelfDeploy} alt="modal-self-deploy " />
-                                                            <h2>Open Source</h2>
-                                                        </div>
-                                                        <h4>Want to use our free open source product?</h4>
-                                                        <p>
-                                                            Deploy PostHog Open Source on your own infrastructure. Free
-                                                            forever.
-                                                        </p>
-                                                    </div>
-                                                </Link>
-                                                <a href="mailto:sales@posthog.com">
-                                                    <div className="modalSelfDeploy modalCard">
-                                                        <div className="modalCardHeader">
-                                                            <img src={modalSelfDeploy} alt="modal-self-deploy " />
-                                                            <h2>Enterprise</h2>
-                                                        </div>
-                                                        <h4>10K+ users? Need support?</h4>
-                                                        <p>
-                                                            Managed on your infrastructure with greater scalability and
-                                                            support. Book a pilot.
-                                                        </p>
-                                                    </div>
-                                                </a>
-                                            </div>
+                                            <GetStartedModalContent />
                                             <Button
                                                 icon="close"
                                                 onClick={() => setModalIsOpen(false)}
