+++ src/pages/req.mdx
@@ -1,0 +1,22 @@
+import Layout from '../components/Layout'
+
+export default Layout
+
+<span class='centered'>
+
+# Request Demo
+
+<div style='max-width: 60vw; margin: auto' class='centered'>
+
+Request a demo to get on a call with one of us! 
+We will do a quick product tour to help you get an in-depth and hands-on overview of the product. 
+Afterwards, we'll answer any questions you might have and help you get started!
+
+</div>
+
+</span>
+
+
+
+
+
