+++ src/pages/startups.js
@@ -13,121 +13,11 @@ import trends from '../images/product-screenshots/product-trends.png'
 import benHorowitz from '../images/ben-horowitz.jpg'
 import michaelSeibel from '../images/michael-seibel.jpg'
 import paulGraham from '../images/paul-graham.jpg'
-import { Col, Row, Form, Input, Card, Tabs } from 'antd'
+import { Col, Row, Card, Tabs } from 'antd'
 import SEO from '../components/seo'
 
 const { TabPane } = Tabs
 const { Meta } = Card
-const layout = {
-    labelCol: { offset: 4, span: 6 },
-    wrapperCol: { span: 6 },
-}
-const tailLayout = {
-    wrapperCol: { offset: 4, span: 16 },
-}
-
-var data = {}
-
-function callback(key) {
-    console.log(key)
-}
-
-function submitForm(e) {
-    e.preventDefault()
-
-    data['form'] = 'startup deal application'
-    data['firstName'] = document.getElementById('firstName').value
-    data['lastName'] = document.getElementById('lastName').value
-    data['email'] = document.getElementById('email').value
-    data['company'] = document.getElementById('company').value
-    data['accelerator'] = document.getElementById('accelerator').value
-    data['deal'] = document.getElementById('deal').value
-
-    var xhr = new XMLHttpRequest()
-    xhr.open('POST', 'https://8qyzxcmhxf.execute-api.us-east-1.amazonaws.com/Prod/submitForm', true)
-    xhr.setRequestHeader('Accept', 'application/json; charset=utf-8')
-    xhr.setRequestHeader('Content-Type', 'application/json; charset=UTF-8')
-    xhr.send(JSON.stringify(data))
-
-    xhr.onloadend = (response) => {
-        if (response.target.status === 200) {
-            window.location.href = '/application-received'
-        } else {
-            alert('error, please email hey@posthog.com')
-            console.error(JSON.parse(response.target.response).message)
-        }
-    }
-}
-
-export function StartupForm({ title, startup = true, subtitle, onSubmit }) {
-    return [
-        <Col className="header-row" align="middle" key="startup-form-1">
-            <h2 id="apply_section">{title}</h2>
-            {subtitle && <p className="subtitle">{subtitle}</p>}
-        </Col>,
-        <Col align="middle" key="startup-form-2">
-            <Form {...layout} name="posthogapplication" id="startupapplication" method="post" onSubmit={onSubmit}>
-                <Form.Item
-                    label="First Name"
-                    name="firstName"
-                    rules={[{ required: true, message: 'Please input your first name!' }]}
-                >
-                    <Input id="firstName" required />
-                </Form.Item>
-                <Form.Item
-                    label="Last Name"
-                    name="lastName"
-                    rules={[{ required: true, message: 'Please input your last name!' }]}
-                >
-                    <Input id="lastName" required />
-                </Form.Item>
-                <Form.Item label="Email" name="email" rules={[{ required: true, message: 'Please input your email!' }]}>
-                    <Input id="email" required type="email" />
-                </Form.Item>
-                <Form.Item
-                    label="Company"
-                    name="company"
-                    rules={[{ required: true, message: 'Please input your company!' }]}
-                >
-                    <Input id="company" required />
-                </Form.Item>
-                {startup && (
-                    <Form.Item
-                        label="Accelerator"
-                        name="accelerator"
-                        rules={[
-                            {
-                                required: true,
-                                message: 'Please provide an accelerator program!',
-                            },
-                        ]}
-                    >
-                        <Input id="accelerator" required />
-                    </Form.Item>
-                )}
-                {startup && (
-                    <Form.Item
-                        label="Deal"
-                        name="deal"
-                        rules={[
-                            {
-                                required: true,
-                                message: 'Please select which deal you would like',
-                            },
-                        ]}
-                    >
-                        <Input id="deal" required />
-                    </Form.Item>
-                )}
-                <Form.Item {...tailLayout}>
-                    <Button type="primary" htmlType="submit">
-                        Submit
-                    </Button>
-                </Form.Item>
-            </Form>
-        </Col>,
-    ]
-}
 
 const StartupsPage = () => {
     useEffect(() => {
@@ -227,7 +117,7 @@ const StartupsPage = () => {
                 </Col>
             </Row>
             <Row gutter={[16, 96]}>
-                <Tabs onChange={callback} type="card">
+                <Tabs type="card">
                     <TabPane tab="Option 1: 🏢 Supported Self-Deploy" key="1">
                         <div className="site-card-wrapper">
                             <Row gutter={16}>
@@ -285,9 +175,6 @@ const StartupsPage = () => {
                         </div>
                     </TabPane>
                 </Tabs>
-            </Row>
-            <Row justify="center">
-                <StartupForm title={'Apply Now'} onSubmit={submitForm}></StartupForm>
             </Row>
             <Row className="gutter-row" justify="center" align="middle">
                 <Col span={24} className="gutter-row" justify="center" align="middle">
